examples/muagent_examples/repochat/requirements.txt
python-dotenv
gitpython
codefuse-muagent 
chainlit